refactor: Phase 2 대형 컴포넌트 분할 완료
ChatInterface.tsx 분할 (1,129줄 → 481줄):
- types.ts: 타입 정의 분리
- styles.ts: 스타일 정의 분리
- hooks/useMisoStreaming.ts: MISO 스트리밍 로직 분리
- hooks/useMentionSystem.ts: 멘션 시스템 로직 분리
- components/ChatHeader.tsx: 헤더 컴포넌트 분리
- components/ChatMessage.tsx: 메시지 컴포넌트 분리
- components/ChatMessages.tsx: 메시지 리스트 컴포넌트 분리
- components/ChatInput.tsx: 입력 컴포넌트 분리
- components/SummaryButton.tsx: 요약 버튼 컴포넌트 분리

ProjectPageContent.tsx 분할 (730줄 → 166줄):
- components/ProjectCard.tsx: 프로젝트 카드 컴포넌트 분리
- components/CreateProjectDialog.tsx: 프로젝트 생성 다이얼로그 분리
- components/ProjectSearchBar.tsx: 검색바 컴포넌트 분리
- components/ProjectSkeleton.tsx: 로딩 스켈레톤 분리
- sections/ProjectHeader.tsx: 헤더 섹션 분리
- sections/ProjectGrid.tsx: 프로젝트 그리드 분리

이제 각 파일이 단일 책임 원칙을 준수하며 100-200줄 이내로 관리됩니다.
